pyproject.toml
[project]
name = "django_utils_kit"
version = "1.0.0"
requires-python = ">=3.9"
description = "Bundle of useful classes and functions for Django"
dependencies = [
    "django>=4.0.0",
    "djangorestframework>=3.13.0",
    "pillow>=11.0.0",
]
readme = "README.md"
authors = [{ name = "Jordan Kowal", email = "kowaljordan@gmail.com" }]
maintainers = [{ name = "Jordan Kowal", email = "kowaljordan@gmail.com" }]
keywords = ["django", "utils", "utilities", "kit", "djangorestframework", "drf"]
license = { file = "LICENSE" }
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX",
    "Operating System :: Unix",
    "Operating System :: MacOS",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Software Development",
]

[project.urls]
Changelog = "https://github.com/Jordan-Kowal/django-utils-kit/blob/main/CHANGELOG.md"
Issues = "https://github.com/Jordan-Kowal/django-utils-kit/issues"
"Release notes" = "https://github.com/Jordan-Kowal/django-utils-kit/releases"
Source = "https://github.com/Jordan-Kowal/django-utils-kit"

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["django_utils_kit"]

# ------------------------------
# Dependencies
# ------------------------------
[tool.uv]
dev-dependencies = [
    "build>=1.2.2.post1",
    "coverage>=7.6.4",
    "mypy>=1.12.1",
    "ruff>=0.7.0",
    "twine>=5.1.1",
]

# ------------------------------
# MYPY
# ------------------------------
[tool.mypy]
python_version = "3.12"
mypy_path = "."

check_untyped_defs = true
disallow_untyped_defs = true
follow_imports = "silent"
ignore_missing_imports = false
show_column_numbers = true
strict_optional = false
warn_redundant_casts = true
warn_unused_ignores = true
disable_error_code = "attr-defined,import-untyped"

# ------------------------------
# RUFF
# ------------------------------
[tool.ruff.format]
quote-style = "double"
docstring-code-format = true

[tool.ruff.lint.isort]
known-first-party = ["django_utils_kit"]
force-sort-within-sections = true

# ------------------------------
# COVERAGE
# ------------------------------
[tool.coverage.run]
# include = ["django_utils_kit/tests/*"]
omit = ["*/__init__.py", "*/tests/*", "*/setup.py"]

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "if TYPE_CHECKING:"]
